+++ DTCD-StyleSystem/src/base-components/components.js
@@ -20,6 +20,7 @@ import BaseExpander from './BaseExpander/BaseExpander';
 import BaseExpanderGroup from './BaseExpanderGroup/BaseExpanderGroup';
 import BaseHeading from './BaseHeading/BaseHeading';
 import BaseFileLoader from './BaseFileLoader/BaseFileLoader';
+import BaseDropdown from './BaseDropdown/BaseDropdown';
 
 export default [
   { name: 'base-button', baseClass: BaseButton },
@@ -44,4 +45,5 @@ export default [
   { name: 'base-file-loader', baseClass: BaseFileLoader },
   { name: 'datasource-select', baseClass: DataSourceSelect },
   { name: 'gauge-segment-builder', baseClass: GaugeSegmentBuilder },
+  { name: 'base-dropdown', baseClass: BaseDropdown },
 ];