+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Label in BaseColorPicker
 - documentation of base components to Storybook
 - BaseFileLoader component
+- generic dropdown list (BaseDropdown)
+- jest config
 
 ### Changed